Split registry with OCI layout
