+++ src/test/java/land/oras/RegistryTest.java
@@ -37,8 +37,6 @@ import java.util.Map;
 import java.util.Random;
 import land.oras.exception.OrasException;
 import land.oras.utils.Const;
-import land.oras.utils.DigestUtils;
-import land.oras.utils.JsonUtils;
 import land.oras.utils.RegistryContainer;
 import land.oras.utils.SupportedAlgorithm;
 import land.oras.utils.ZotContainer;
@@ -58,17 +56,8 @@ public class RegistryTest {
     @Container
     private final ZotContainer registry = new ZotContainer().withStartupAttempts(3);
 
-    /**
-     * Blob temporary dir
-     */
     @TempDir
     private Path blobDir;
-
-    @TempDir
-    private Path ociLayout;
-
-    @TempDir
-    private Path ociLayoutWithIndex;
 
     @TempDir
     private Path artifactDir;
@@ -509,55 +498,6 @@ public class RegistryTest {
     }
 
     @Test
-    void testShouldCopyArtifactIntoOciLayout() throws IOException {
-        Registry registry = Registry.Builder.builder()
-                .defaults("myuser", "mypass")
-                .withInsecure(true)
-                .build();
-
-        ContainerRef containerRef =
-                ContainerRef.parse("%s/library/artifact-oci-layout".formatted(this.registry.getRegistry()));
-        Path file1 = blobDir.resolve("artifact-oci-layout.txt");
-        Files.writeString(file1, "artifact-oci-layout");
-
-        // Push
-        Manifest manifest = registry.pushArtifact(containerRef, LocalPath.of(file1));
-
-        // Cannot copy if directory doesn't exists
-        assertThrows(
-                OrasException.class,
-                () -> {
-                    registry.copy(containerRef, ociLayout.resolve("not-exists"));
-                },
-                "Directory not found");
-
-        // Copy to oci layout
-        registry.copy(containerRef, ociLayout);
-
-        assertTrue(Files.exists(ociLayout.resolve("oci-layout")));
-
-        OciLayout layoutFile = JsonUtils.fromJson(ociLayout.resolve("oci-layout"), OciLayout.class);
-        assertEquals("1.0.0", layoutFile.getImageLayoutVersion());
-
-        // Assert the empty config
-        assertEquals(
-                "{}",
-                Files.readString(ociLayout
-                        .resolve("blobs")
-                        .resolve("sha256")
-                        .resolve(SupportedAlgorithm.getDigest(Config.empty().getDigest()))));
-
-        // Check index exists
-        assertTrue(Files.exists(ociLayout.resolve("index.json")));
-        Index index = JsonUtils.fromJson(ociLayout.resolve("index.json"), Index.class);
-        assertEquals(2, index.getSchemaVersion());
-        assertEquals(1, index.getManifests().size());
-        assertEquals(Const.DEFAULT_INDEX_MEDIA_TYPE, index.getMediaType());
-        assertEquals(
-                manifest.getDescriptor().getSize(), index.getManifests().get(0).getSize());
-    }
-
-    @Test
     void testNotFailToPullArtifactFromImage() {
 
         Registry registry = Registry.Builder.builder()
@@ -589,217 +529,6 @@ public class RegistryTest {
         assertEquals(1, pushIndex.getManifests().size());
         assertEquals(2, pushIndex.getSchemaVersion());
         assertEquals(Const.DEFAULT_INDEX_MEDIA_TYPE, pushIndex.getMediaType());
-    }
-
-    @Test
-    void testShouldCopyImageIntoOciLayoutWithoutIndex() {
-
-        Registry registry = Registry.Builder.builder()
-                .defaults("myuser", "mypass")
-                .withInsecure(true)
-                .build();
-
-        ContainerRef containerRef =
-                ContainerRef.parse("%s/library/image-no-index".formatted(this.registry.getRegistry()));
-
-        Layer layer1 = registry.pushBlob(containerRef, Layer.empty().getDataBytes());
-        Layer layer2 = registry.pushBlob(containerRef, "foobar".getBytes());
-
-        Manifest emptyManifest = Manifest.empty()
-                .withLayers(List.of(Layer.fromDigest(layer1.getDigest(), 2), Layer.fromDigest(layer2.getDigest(), 6)));
-        String manifestDigest =
-                SupportedAlgorithm.SHA256.digest(emptyManifest.toJson().getBytes(StandardCharsets.UTF_8));
-        String configDigest =
-                SupportedAlgorithm.SHA256.digest(Config.empty().toJson().getBytes(StandardCharsets.UTF_8));
-
-        // Push config and manifest
-        registry.pushConfig(containerRef.withDigest(configDigest), Config.empty());
-        Manifest pushedManifest = registry.pushManifest(containerRef, emptyManifest);
-
-        // Copy to oci layout
-        registry.copy(containerRef, ociLayout);
-
-        assertTrue(Files.exists(ociLayout.resolve("oci-layout")));
-
-        OciLayout layoutFile = JsonUtils.fromJson(ociLayout.resolve("oci-layout"), OciLayout.class);
-        assertEquals("1.0.0", layoutFile.getImageLayoutVersion());
-
-        // Check index exists
-        assertTrue(Files.exists(ociLayout.resolve("index.json")));
-        JsonUtils.fromJson(ociLayout.resolve("index.json"), Index.class);
-
-        // Check manifest exists
-        assertTrue(Files.exists(ociLayout
-                .resolve("blobs")
-                .resolve("sha256")
-                .resolve(SupportedAlgorithm.getDigest(
-                        pushedManifest.getDescriptor().getDigest()))));
-
-        // Ensure manifest serialized correctly (check sha256)
-        String computedManifestDigest = DigestUtils.digest(
-                "sha256",
-                ociLayout
-                        .resolve("blobs")
-                        .resolve("sha256")
-                        .resolve(SupportedAlgorithm.getDigest(
-                                pushedManifest.getDescriptor().getDigest())));
-        assertEquals(
-                SupportedAlgorithm.getDigest(pushedManifest.getDescriptor().getDigest()),
-                SupportedAlgorithm.getDigest(computedManifestDigest),
-                "Manifest digest should match");
-
-        // Ensure layer1 is copied
-        assertTrue(Files.exists(ociLayout
-                .resolve("blobs")
-                .resolve("sha256")
-                .resolve(SupportedAlgorithm.getDigest(layer1.getDigest()))));
-        // Ensure layer2 is copied
-        assertTrue(Files.exists(ociLayout
-                .resolve("blobs")
-                .resolve("sha256")
-                .resolve(SupportedAlgorithm.getDigest(layer2.getDigest()))));
-
-        // Copy to oci layout again
-        registry.copy(containerRef, ociLayout);
-
-        // Check manifest exists
-        assertTrue(Files.exists(ociLayout
-                .resolve("blobs")
-                .resolve("sha256")
-                .resolve(SupportedAlgorithm.getDigest(
-                        pushedManifest.getDescriptor().getDigest()))));
-    }
-
-    @Test
-    void testShouldCopyImageIntoOciLayoutWithIndex() {
-
-        Registry registry = Registry.Builder.builder()
-                .defaults("myuser", "mypass")
-                .withInsecure(true)
-                .build();
-
-        ContainerRef containerRef =
-                ContainerRef.parse("%s/library/artifact-image-pull".formatted(this.registry.getRegistry()));
-
-        Layer layer1 = registry.pushBlob(containerRef, Layer.empty().getDataBytes());
-        Layer layer2 = registry.pushBlob(containerRef, "foobar".getBytes());
-
-        Manifest emptyManifest = Manifest.empty()
-                .withLayers(List.of(Layer.fromDigest(layer1.getDigest(), 2), Layer.fromDigest(layer2.getDigest(), 6)));
-        String manifestDigest =
-                SupportedAlgorithm.SHA256.digest(emptyManifest.toJson().getBytes(StandardCharsets.UTF_8));
-        String configDigest =
-                SupportedAlgorithm.SHA256.digest(Config.empty().toJson().getBytes(StandardCharsets.UTF_8));
-
-        // Push config and manifest
-        registry.pushConfig(containerRef.withDigest(configDigest), Config.empty());
-        Manifest pushedManifest = registry.pushManifest(containerRef.withDigest(manifestDigest), emptyManifest);
-        Index index = registry.pushIndex(containerRef, Index.fromManifests(List.of(pushedManifest.getDescriptor())));
-
-        // Copy to oci layout
-        registry.copy(containerRef, ociLayoutWithIndex);
-
-        assertTrue(Files.exists(ociLayoutWithIndex.resolve("oci-layout")));
-
-        OciLayout layoutFile = JsonUtils.fromJson(ociLayoutWithIndex.resolve("oci-layout"), OciLayout.class);
-        assertEquals("1.0.0", layoutFile.getImageLayoutVersion());
-
-        // Check index exists
-        assertTrue(Files.exists(ociLayoutWithIndex.resolve("index.json")));
-        JsonUtils.fromJson(ociLayoutWithIndex.resolve("index.json"), Index.class);
-        assertEquals(2, index.getSchemaVersion());
-        assertEquals(1, index.getManifests().size());
-        assertEquals(Const.DEFAULT_INDEX_MEDIA_TYPE, index.getMediaType());
-
-        // Check manifest exists
-        assertTrue(Files.exists(ociLayoutWithIndex
-                .resolve("blobs")
-                .resolve("sha256")
-                .resolve(SupportedAlgorithm.getDigest(
-                        pushedManifest.getDescriptor().getDigest()))));
-
-        // Ensure manifest serialized correctly (check sha256)
-        String computedManifestDigest = DigestUtils.digest(
-                "sha256",
-                ociLayoutWithIndex
-                        .resolve("blobs")
-                        .resolve("sha256")
-                        .resolve(SupportedAlgorithm.getDigest(
-                                pushedManifest.getDescriptor().getDigest())));
-        assertEquals(
-                SupportedAlgorithm.getDigest(pushedManifest.getDescriptor().getDigest()),
-                SupportedAlgorithm.getDigest(computedManifestDigest),
-                "Manifest digest should match");
-
-        // Ensure layer1 is copied
-        assertTrue(Files.exists(ociLayoutWithIndex
-                .resolve("blobs")
-                .resolve("sha256")
-                .resolve(SupportedAlgorithm.getDigest(layer1.getDigest()))));
-        // Ensure layer2 is copied
-        assertTrue(Files.exists(ociLayoutWithIndex
-                .resolve("blobs")
-                .resolve("sha256")
-                .resolve(SupportedAlgorithm.getDigest(layer2.getDigest()))));
-        // Ensure index is also copied as blob
-        assertTrue(Files.exists(ociLayoutWithIndex
-                .resolve("blobs")
-                .resolve("sha256")
-                .resolve(SupportedAlgorithm.getDigest(index.getDescriptor().getDigest()))));
-
-        // Copy to oci layout again
-        registry.copy(containerRef, ociLayoutWithIndex);
-
-        // Check manifest exists
-        assertTrue(Files.exists(ociLayoutWithIndex
-                .resolve("blobs")
-                .resolve("sha256")
-                .resolve(SupportedAlgorithm.getDigest(
-                        pushedManifest.getDescriptor().getDigest()))));
-    }
-
-    @Test
-    void testShouldCopyIntoOciLayoutWithBlobConfig() throws IOException {
-        Registry registry = Registry.Builder.builder()
-                .defaults("myuser", "mypass")
-                .withInsecure(true)
-                .build();
-
-        ContainerRef containerRef =
-                ContainerRef.parse("%s/library/artifact-oci-layout".formatted(this.registry.getRegistry()));
-        Path file1 = blobDir.resolve("artifact-oci-layout.txt");
-        Files.writeString(file1, "artifact-oci-layout");
-
-        // Push
-        Layer layer = registry.pushBlob(containerRef, "foobartest".getBytes(StandardCharsets.UTF_8));
-        Config config = Config.fromBlob("text/plain", layer);
-        Manifest manifest = registry.pushArtifact(
-                containerRef, ArtifactType.from("my/artifact"), Annotations.empty(), config, LocalPath.of(file1));
-
-        // Copy to oci layout
-        registry.copy(containerRef, ociLayout);
-
-        assertTrue(Files.exists(ociLayout.resolve("oci-layout")));
-
-        OciLayout layoutFile = JsonUtils.fromJson(ociLayout.resolve("oci-layout"), OciLayout.class);
-        assertEquals("1.0.0", layoutFile.getImageLayoutVersion());
-
-        // Assert the config
-        assertEquals(
-                "foobartest",
-                Files.readString(ociLayout
-                        .resolve("blobs")
-                        .resolve("sha256")
-                        .resolve(SupportedAlgorithm.getDigest(layer.getDigest()))));
-
-        // Check index exists
-        assertTrue(Files.exists(ociLayout.resolve("index.json")));
-        Index index = JsonUtils.fromJson(ociLayout.resolve("index.json"), Index.class);
-        assertEquals(2, index.getSchemaVersion());
-        assertEquals(1, index.getManifests().size());
-        assertEquals(Const.DEFAULT_INDEX_MEDIA_TYPE, index.getMediaType());
-        assertEquals(
-                manifest.getDescriptor().getSize(), index.getManifests().get(0).getSize());
     }
 
     @Test
